split 9k line file into multiple smaller
