+++ src/System.Private.Xml/tests/Writers/XmlWriterApi/System.Xml.RW.XmlWriterApi.Tests.csproj
@@ -13,17 +13,23 @@
     <ReferenceFromRuntime Include="xunit.execution.dotnet" />
   </ItemGroup>
   <ItemGroup>
-    <!--<Compile Include="CommonTests.cs" />-->
     <Compile Include="EndOfLineHandlingTests.cs" />
     <Compile Include="ErrorCondition.cs" />
     <Compile Include="NamespaceHandlingTests.cs" />
-    <!--<Compile Include="ReaderParamTestCase.cs" />-->
+    <!--<Compile Include="TCAttribute.cs" />-->
+    <!--<Compile Include="ReaderParamTestCase.cs" /> ???? -->
     <Compile Include="TCAutoCL.cs" />
+    <!--<Compile Include="TCAutoComplete.cs" />-->
     <Compile Include="TCCheckChars.cs" />
     <Compile Include="TCCloseOutput.cs" />
     <Compile Include="TCDefaultWriterSettingsTests.cs" />
+    <!--<Compile Include="TCDocType.cs" />
+    <Compile Include="TCDocument.cs" />
+    <Compile Include="TCElement.cs" />
+    <Compile Include="TCErrorState.cs" />-->
     <Compile Include="TCFlushClose.cs" />
     <Compile Include="TCFragmentCL.cs" />
+    <!--<Compile Include="TCFullEndElement.cs" />-->
     <Compile Include="TCIndent.cs" />
     <Compile Include="TCIndentChars.cs" />
     <Compile Include="TCNewLineChars.cs" />
@@ -32,6 +38,9 @@
     <Compile Include="DisableParallelization.cs" />
     <Compile Include="TCOmitXmlDecl.cs" />
     <Compile Include="TCStandAlone.cs" />
+    <!--<Compile Include="TCWriteAttributes.cs" />
+    <Compile Include="TCWriteNode_With_ReadValueChunk.cs" />
+    <Compile Include="TCWriteNode_XmlReader.cs" />-->
     <Compile Include="TCWriterSettingsMisc.cs" />
     <Compile Include="TCWriterWithMemoryStream.cs" />
     <Compile Include="TestExtensions.cs" />
