+++ src/System.Private.Xml/tests/Writers/XmlWriterApi/TCErrorState.cs
@@ -1,0 +1,315 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using OLEDB.Test.ModuleCore;
+using XmlCoreTest.Common;
+using Xunit;
+
+namespace System.Xml.Tests
+{
+    //[TestCase(Name = "Invalid State Combinations", Pri = 1)]
+    public class TCErrorState : XmlWriterTestCaseBase
+    {
+        //[Variation(id = 1, Desc = "EntityRef after Document should error - PROLOG", Pri = 1)]
+        [Fact]
+        public void state_1()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartDocument();
+                    w.WriteEntityRef("ent");
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+        //[Variation(id = 2, Desc = "EntityRef after Document should error - EPILOG", Pri = 1)]
+        [Fact]
+        public void state_2()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartDocument();
+                    w.WriteStartElement("Root");
+                    w.WriteEndElement();
+                    w.WriteEntityRef("ent");
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+        //[Variation(id = 3, Desc = "CharEntity after Document should error - PROLOG", Pri = 1)]
+        [Fact]
+        public void state_3()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartDocument();
+                    w.WriteCharEntity('\uD23E');
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+        //[Variation(id = 4, Desc = "CharEntity after Document should error - EPILOG", Pri = 1)]
+        [Fact]
+        public void state_4()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartDocument();
+                    w.WriteStartElement("Root");
+                    w.WriteEndElement();
+                    w.WriteCharEntity('\uD23E');
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+        //[Variation(id = 5, Desc = "SurrogateCharEntity after Document should error - PROLOG", Pri = 1)]
+        [Fact]
+        public void state_5()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartDocument();
+                    w.WriteSurrogateCharEntity('\uDF41', '\uD920');
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+        //[Variation(id = 6, Desc = "SurrogateCharEntity after Document should error - EPILOG", Pri = 1)]
+        [Fact]
+        public void state_6()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartDocument();
+                    w.WriteStartElement("Root");
+                    w.WriteEndElement();
+                    w.WriteSurrogateCharEntity('\uDF41', '\uD920');
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+        //[Variation(id = 7, Desc = "Attribute after Document should error - PROLOG", Pri = 1)]
+        [Fact]
+        public void state_7()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartDocument();
+                    w.WriteStartAttribute("attr", "");
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+        //[Variation(id = 8, Desc = "Attribute after Document should error - EPILOG", Pri = 1)]
+        [Fact]
+        public void state_8()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartDocument();
+                    w.WriteStartElement("Root");
+                    w.WriteEndElement();
+                    w.WriteStartAttribute("attr", "");
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+        //[Variation(id = 9, Desc = "CDATA after Document should error - PROLOG", Pri = 1)]
+        [Fact]
+        public void state_9()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartDocument();
+                    w.WriteCData("Invalid");
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            return;
+        }
+
+        //[Variation(id = 10, Desc = "CDATA after Document should error - EPILOG", Pri = 1)]
+        [Fact]
+        public void state_10()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartDocument();
+                    w.WriteStartElement("Root");
+                    w.WriteEndElement();
+                    w.WriteCData("Invalid");
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+        //[Variation(id = 11, Desc = "Element followed by Document should error", Pri = 1)]
+        [Fact]
+        public void state_11()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartElement("Root");
+                    w.WriteStartDocument();
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+        //[Variation(id = 12, Desc = "Element followed by DocType should error", Pri = 1)]
+        [Fact]
+        public void state_12()
+        {
+            using (XmlWriter w = CreateWriter())
+            {
+                try
+                {
+                    w.WriteStartElement("Root");
+                    w.WriteDocType("Test", null, null, "");
+                }
+                catch (InvalidOperationException e)
+                {
+                    CError.WriteLineIgnore("Exception: " + e.ToString());
+                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
+                    return;
+                }
+            }
+            CError.WriteLine("Did not throw exception");
+            Assert.True(false);
+        }
+
+
+        //[Variation(Desc = "1.WriteNode with GB18030 encoding", Param = 1)]
+        //[Variation(Desc = "2.WriteNode with GB18030 encoding", Param = 2)]
+        [Fact]
+        public void writeNode_XmlReader36a()
+        {
+            string path = FilePathUtil.GetStandardPath();
+            int param = (int)this.CurVariation.Param;
+            string xml = (param == 1) ? path + @"\xml10\ms_xml\bug433100.xml" : path + @"\Globalization\Lang\GB18030\gb18030char.xml";
+            CError.WriteLine(xml);
+
+            XmlWriterSettings ws = new XmlWriterSettings();
+            ws.OmitXmlDeclaration = true;
+            using (XmlReader xr = ReaderHelper.Create(xml))
+            {
+                using (XmlWriter w = CreateWriter(ws))
+                {
+                    while (!xr.EOF)
+                    {
+                        w.WriteNode(xr, true);
+                    }
+                }
+            }
+            Assert.True(CompareBaseline2(xml));
+        }
+    }
+}
