+++ src/System.Private.Xml/tests/Writers/XmlWriterApi/TCFullEndElement.cs
@@ -12,3405 +12,8 @@ using Xunit;
 
 namespace System.Xml.Tests
 {
-    //[TestCase(Name = "Invalid State Combinations", Pri = 1)]
-    public partial class TCErrorState : XmlWriterTestCaseBase
-    {
-        //[Variation(id = 1, Desc = "EntityRef after Document should error - PROLOG", Pri = 1)]
-        [Fact]
-        public void state_1()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteEntityRef("ent");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 2, Desc = "EntityRef after Document should error - EPILOG", Pri = 1)]
-        [Fact]
-        public void state_2()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteEntityRef("ent");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 3, Desc = "CharEntity after Document should error - PROLOG", Pri = 1)]
-        [Fact]
-        public void state_3()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteCharEntity('\uD23E');
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 4, Desc = "CharEntity after Document should error - EPILOG", Pri = 1)]
-        [Fact]
-        public void state_4()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteCharEntity('\uD23E');
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 5, Desc = "SurrogateCharEntity after Document should error - PROLOG", Pri = 1)]
-        [Fact]
-        public void state_5()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteSurrogateCharEntity('\uDF41', '\uD920');
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 6, Desc = "SurrogateCharEntity after Document should error - EPILOG", Pri = 1)]
-        [Fact]
-        public void state_6()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteSurrogateCharEntity('\uDF41', '\uD920');
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 7, Desc = "Attribute after Document should error - PROLOG", Pri = 1)]
-        [Fact]
-        public void state_7()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartAttribute("attr", "");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 8, Desc = "Attribute after Document should error - EPILOG", Pri = 1)]
-        [Fact]
-        public void state_8()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteStartAttribute("attr", "");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 9, Desc = "CDATA after Document should error - PROLOG", Pri = 1)]
-        [Fact]
-        public void state_9()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteCData("Invalid");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            return;
-        }
-
-        //[Variation(id = 10, Desc = "CDATA after Document should error - EPILOG", Pri = 1)]
-        [Fact]
-        public void state_10()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteCData("Invalid");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 11, Desc = "Element followed by Document should error", Pri = 1)]
-        [Fact]
-        public void state_11()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteStartDocument();
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 12, Desc = "Element followed by DocType should error", Pri = 1)]
-        [Fact]
-        public void state_12()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteDocType("Test", null, null, "");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-
-        //[Variation(Desc = "1.WriteNode with GB18030 encoding", Param = 1)]
-        //[Variation(Desc = "2.WriteNode with GB18030 encoding", Param = 2)]
-        [Fact]
-        public void writeNode_XmlReader36a()
-        {
-            string path = FilePathUtil.GetStandardPath();
-            int param = (int)this.CurVariation.Param;
-            string xml = (param == 1) ? path + @"\xml10\ms_xml\bug433100.xml" : path + @"\Globalization\Lang\GB18030\gb18030char.xml";
-            CError.WriteLine(xml);
-
-            XmlWriterSettings ws = new XmlWriterSettings();
-            ws.OmitXmlDeclaration = true;
-            using (XmlReader xr = ReaderHelper.Create(xml))
-            {
-                using (XmlWriter w = CreateWriter(ws))
-                {
-                    while (!xr.EOF)
-                    {
-                        w.WriteNode(xr, true);
-                    }
-                }
-            }
-            Assert.True(CompareBaseline2(xml));
-        }
-    }
-
-    ////[TestCase(Name = "Auto-completion of tokens")]
-    public partial class TCAutoComplete : XmlWriterTestCaseBase
-    {
-        //[Variation(id = 1, Desc = "Missing EndAttr, followed by element", Pri = 1)]
-        [Fact]
-        public void var_1()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteStartAttribute("attr");
-                w.WriteStartElement("child");
-                w.WriteEndElement();
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root attr=''><child /></Root>"));
-        }
-
-        //[Variation(id = 2, Desc = "Missing EndAttr, followed by comment", Pri = 1)]
-        [Fact]
-        public void var_2()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteStartAttribute("attr");
-                w.WriteComment("This text is a comment");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root attr=''><!--This text is a comment--></Root>"));
-        }
-
-        //[Variation(id = 3, Desc = "Write EndDocument with unclosed element tag", Pri = 1)]
-        [Fact]
-        public void var_3()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartDocument();
-                w.WriteStartElement("Root");
-                w.WriteEndDocument();
-            }
-            Assert.True(CompareReader("<Root />"));
-        }
-
-        //[Variation(id = 4, Desc = "WriteStartDocument - WriteEndDocument", Pri = 1)]
-        [Fact]
-        public void var_4()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteEndDocument();
-                }
-                catch (ArgumentException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-
-                    Assert.True(false);
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 5, Desc = "WriteEndElement without WriteStartElement", Pri = 1)]
-        [Fact]
-        public void var_5()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteEndElement();
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 6, Desc = "WriteFullEndElement without WriteStartElement", Pri = 1)]
-        [Fact]
-        public void var_6()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteFullEndElement();
-                    w.WriteFullEndElement();
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-    }
-
-    //[TestCase(Name = "WriteStart/EndDocument")]
-    public partial class TCDocument : XmlWriterTestCaseBase
-    {
-        //[Variation(id = 1, Desc = "StartDocument-EndDocument Sanity Test", Pri = 0)]
-        [Fact]
-        public void document_1()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartDocument();
-                w.WriteStartElement("Root");
-                w.WriteEndElement();
-                w.WriteEndDocument();
-            }
-            Assert.True(CompareReader("<Root />"));
-        }
-
-        //[Variation(id = 2, Desc = "Multiple StartDocument should error", Pri = 1)]
-        [Fact]
-        public void document_2()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartDocument();
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 3, Desc = "Missing StartDocument should be fixed", Pri = 1)]
-        [Fact]
-        public void document_3()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteEndElement();
-                w.WriteEndDocument();
-            }
-            Assert.True(CompareReader("<Root />"));
-        }
-
-
-        //[Variation(id = 4, Desc = "Multiple EndDocument should error", Pri = 1)]
-        [Fact]
-        public void document_4()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteEndDocument();
-                    w.WriteEndDocument();
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 5, Desc = "Missing EndDocument should be fixed", Pri = 1)]
-        [Fact]
-        public void document_5()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartDocument();
-                w.WriteStartElement("Root");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root />"));
-        }
-
-        //[Variation(id = 6, Desc = "Call Start-EndDocument multiple times, should error", Pri = 2)]
-        [Fact]
-        public void document_6()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteEndDocument();
-
-                    w.WriteStartDocument();
-                    w.WriteEndDocument();
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 7, Desc = "Multiple root elements should error", Pri = 1)]
-        [Fact]
-        public void document_7()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteStartElement("Root");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 8, Desc = "Start-EndDocument without any element should error", Pri = 2)]
-        [Fact]
-        public void document_8()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteEndDocument();
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 9, Desc = "Top level text should error - PROLOG", Pri = 1)]
-        [Fact]
-        public void document_9()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteString("Top level text");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 10, Desc = "Top level text should error - EPILOG", Pri = 1)]
-        [Fact]
-        public void document_10()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteString("Top level text");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-
-        //[Variation(id = 11, Desc = "Top level atomic value should error - PROLOG", Pri = 1)]
-        [Fact]
-        public void document_11()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    int i = 1;
-                    w.WriteValue(i);
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            Assert.True(false);
-        }
-
-        //[Variation(id = 12, Desc = "Top level atomic value should error - EPILOG", Pri = 1)]
-        [Fact]
-        public void document_12()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartDocument();
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    int i = 1;
-                    w.WriteValue(i);
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            Assert.True(false);
-        }
-    }
-
-    //[TestCase(Name = "WriteDocType")]
-    public partial class TCDocType : XmlWriterTestCaseBase
-    {
-        //[Variation(id = 1, Desc = "Sanity test", Pri = 1)]
-        [Fact]
-        public void docType_1()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteDocType("ROOT", "publicid", "sysid", "<!ENTITY e 'abc'>");
-                w.WriteStartElement("ROOT");
-                w.WriteEndElement();
-            }
-
-            string exp = IsIndent() ?
-                "<!DOCTYPE ROOT PUBLIC \"publicid\" \"sysid\"[<!ENTITY e 'abc'>]>" + Environment.NewLine + "<ROOT />" :
-                "<!DOCTYPE ROOT PUBLIC \"publicid\" \"sysid\"[<!ENTITY e 'abc'>]><ROOT />";
-            Assert.True(CompareString(exp));
-        }
-
-        //[Variation(id = 2, Desc = "WriteDocType pubid = null and sysid = null", Pri = 1)]
-        [Fact]
-        public void docType_2()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteDocType("test", null, null, "<!ENTITY e 'abc'>");
-                w.WriteStartElement("Root");
-                w.WriteEndElement();
-            }
-            string exp = IsIndent() ?
-                "<!DOCTYPE test [<!ENTITY e 'abc'>]>" + Environment.NewLine + "<Root />" :
-                "<!DOCTYPE test [<!ENTITY e 'abc'>]><Root />";
-            Assert.True(CompareString(exp));
-        }
-
-        //[Variation(id = 3, Desc = "Call WriteDocType twice", Pri = 1)]
-        [Fact]
-        public void docType_3()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteDocType("doc1", null, null, "<!ENTITY e 'abc'>");
-                    w.WriteDocType("doc2", null, null, "<!ENTITY f 'abc'>");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore(e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 4, Desc = "WriteDocType with name value = String.Empty", Param = "String.Empty", Pri = 1)]
-        //[Variation(id = 5, Desc = "WriteDocType with name value = null", Param = "null", Pri = 1)]
-        [Fact]
-        public void docType_4()
-        {
-            String docName = "";
-            if (CurVariation.Param.ToString() == "String.Empty")
-                docName = String.Empty;
-            else if (CurVariation.Param.ToString() == "null")
-                docName = null;
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteDocType(docName, null, null, "test1");
-                }
-                catch (ArgumentException e)
-                {
-                    CError.WriteLineIgnore(e.ToString());
-                    CError.Compare(w.WriteState, (WriterType == WriterType.CharCheckingWriter) ? WriteState.Start : WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-                catch (NullReferenceException e)
-                {
-                    CError.WriteLineIgnore(e.ToString());
-                    CError.Compare(w.WriteState, (WriterType == WriterType.CharCheckingWriter) ? WriteState.Start : WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 6, Desc = "WriteDocType with DocType end tag in the value", Pri = 1)]
-        [Fact]
-        public void docType_5()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                String docName = "Root";
-                String docValue = "]>";
-                w.WriteDocType(docName, null, null, docValue);
-                w.WriteStartElement("Root");
-                w.WriteEndElement();
-            }
-            string exp = IsIndent() ? "<!DOCTYPE Root []>]>" + Environment.NewLine + "<Root />" : "<!DOCTYPE Root []>]><Root />";
-            Assert.True(CompareString(exp));
-        }
-
-        //[Variation(id = 7, Desc = "Call WriteDocType in the root element", Pri = 1)]
-        [Fact]
-        public void docType_6()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteDocType("doc1", null, null, "test1");
-                    w.WriteEndElement();
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore(e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 8, Desc = "Call WriteDocType following root element", Pri = 1)]
-        [Fact]
-        public void docType_7()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteEndElement();
-                    w.WriteDocType("doc1", null, null, "test1");
-                }
-                catch (InvalidOperationException e)
-                {
-                    CError.WriteLineIgnore(e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-    }
-
-    //[TestCase(Name = "WriteStart/EndElement")]
-    public partial class TCElement : XmlWriterTestCaseBase
-    {
-        //[Variation(id = 1, Desc = "StartElement-EndElement Sanity Test", Pri = 0)]
-        [Fact]
-        public void element_1()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root />"));
-        }
-
-        //[Variation(id = 2, Desc = "Sanity test for overload WriteStartElement(string prefix, string name, string ns)", Pri = 0)]
-        [Fact]
-        public void element_2()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("pre1", "Root", "http://my.com");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<pre1:Root xmlns:pre1=\"http://my.com\" />"));
-        }
-
-        //[Variation(id = 3, Desc = "Sanity test for overload WriteStartElement(string name, string ns)", Pri = 0)]
-        [Fact]
-        public void element_3()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root", "http://my.com");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root xmlns=\"http://my.com\" />"));
-        }
-
-        //[Variation(id = 4, Desc = "Element name = String.Empty should error", Pri = 1)]
-        [Fact]
-        public void element_4()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement(String.Empty);
-                }
-                catch (ArgumentException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, (WriterType == WriterType.CharCheckingWriter) ? WriteState.Start : WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 5, Desc = "Element name = null should error", Pri = 1)]
-        [Fact]
-        public void element_5()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement(null);
-                }
-                catch (ArgumentException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, (WriterType == WriterType.CharCheckingWriter) ? WriteState.Start : WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 6, Desc = "Element NS = String.Empty", Pri = 1)]
-        [Fact]
-        public void element_6()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root", String.Empty);
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root />"));
-        }
-
-        //[Variation(id = 7, Desc = "Element NS = null", Pri = 1)]
-        [Fact]
-        public void element_7()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root", null);
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root />"));
-        }
-
-        //[Variation(Desc = "Write 100 nested elements")]
-        [Fact]
-        public void element_8()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                for (int i = 0; i < 100; i++)
-                {
-                    string eName = "Node" + i.ToString();
-                    w.WriteStartElement(eName);
-                }
-                for (int i = 0; i < 100; i++)
-                    w.WriteEndElement();
-            }
-
-            string exp = (WriterType == WriterType.UTF8WriterIndent || WriterType == WriterType.UnicodeWriterIndent) ?
-                "100ElementsIndent.txt" : "100Elements.txt";
-            Assert.True(CompareBaseline(exp));
-        }
-
-        //[Variation(id = 9, Desc = "WriteDecl with start element with prefix and namespace")]
-        [Fact]
-        public void element_9()
-        {
-            string enc = (WriterType == WriterType.UnicodeWriter || WriterType == WriterType.UnicodeWriterIndent) ? "16" : "8";
-            string exp = (WriterType == WriterType.UTF8WriterIndent || WriterType == WriterType.UnicodeWriterIndent) ?
-                String.Format("<?xml version=\"1.0\" encoding=\"utf-{0}\"?>" + nl + "<a:b xmlns:a=\"c\" />", enc) :
-                String.Format("<?xml version=\"1.0\" encoding=\"utf-{0}\"?><a:b xmlns:a=\"c\" />", enc);
-
-            XmlWriterSettings ws = new XmlWriterSettings();
-            ws.OmitXmlDeclaration = false;
-            using (XmlWriter w = CreateWriter(ws))
-            {
-                w.WriteStartElement("a", "b", "c");
-            }
-
-            Assert.True((CompareString(exp)));
-        }
-
-        //[Variation(Desc = "Write many attributes with same names and diff.namespaces", Param = true)]
-        //[Variation(Desc = "Write many attributes with same names and diff.namespaces", Param = false)]
-        [Fact]
-        public void element_10()
-        {
-            string xml = "<a p1:a=\"\" p2:a=\"\" p3:a=\"\" p4:a=\"\" p5:a=\"\" p6:a=\"\" p7:a=\"\" p8:a=\"\" p9:a=\"\" p10:a=\"\" p11:a=\"\" p12:a=\"\" p13:a=\"\" p14:a=\"\" p15:a=\"\" p16:a=\"\" p17:a=\"\" p18:a=\"\" p19:a=\"\" p20:a=\"\" p21:a=\"\" p22:a=\"\" p23:a=\"\" p24:a=\"\" p25:a=\"\" p26:a=\"\" p27:a=\"\" p28:a=\"\" p29:a=\"\" p30:a=\"\" p31:a=\"\" p32:a=\"\" p33:a=\"\" p34:a=\"\" p35:a=\"\" p36:a=\"\" p37:a=\"\" p38:a=\"\" p39:a=\"\" p40:a=\"\" p41:a=\"\" p42:a=\"\" p43:a=\"\" p44:a=\"\" p45:a=\"\" p46:a=\"\" p47:a=\"\" p48:a=\"\" p49:a=\"\" p50:a=\"\" p51:a=\"\" p52:a=\"\" p53:a=\"\" p54:a=\"\" p55:a=\"\" p56:a=\"\" p57:a=\"\" p58:a=\"\" p59:a=\"\" p60:a=\"\" p61:a=\"\" p62:a=\"\" p63:a=\"\" p64:a=\"\" p65:a=\"\" p66:a=\"\" p67:a=\"\" p68:a=\"\" p69:a=\"\" p70:a=\"\" p71:a=\"\" p72:a=\"\" p73:a=\"\" p74:a=\"\" p75:a=\"\" p76:a=\"\" p77:a=\"\" p78:a=\"\" p79:a=\"\" p80:a=\"\" p81:a=\"\" p82:a=\"\" p83:a=\"\" p84:a=\"\" p85:a=\"\" p86:a=\"\" p87:a=\"\" p88:a=\"\" p89:a=\"\" p90:a=\"\" p91:a=\"\" p92:a=\"\" p93:a=\"\" p94:a=\"\" p95:a=\"\" p96:a=\"\" p97:a=\"\" p98:a=\"\" p99:a=\"\" p100:a=\"\" xmlns:p100=\"b99\" xmlns:p99=\"b98\" xmlns:p98=\"b97\" xmlns:p97=\"b96\" xmlns:p96=\"b95\" xmlns:p95=\"b94\" xmlns:p94=\"b93\" xmlns:p93=\"b92\" xmlns:p92=\"b91\" xmlns:p91=\"b90\" xmlns:p90=\"b89\" xmlns:p89=\"b88\" xmlns:p88=\"b87\" xmlns:p87=\"b86\" xmlns:p86=\"b85\" xmlns:p85=\"b84\" xmlns:p84=\"b83\" xmlns:p83=\"b82\" xmlns:p82=\"b81\" xmlns:p81=\"b80\" xmlns:p80=\"b79\" xmlns:p79=\"b78\" xmlns:p78=\"b77\" xmlns:p77=\"b76\" xmlns:p76=\"b75\" xmlns:p75=\"b74\" xmlns:p74=\"b73\" xmlns:p73=\"b72\" xmlns:p72=\"b71\" xmlns:p71=\"b70\" xmlns:p70=\"b69\" xmlns:p69=\"b68\" xmlns:p68=\"b67\" xmlns:p67=\"b66\" xmlns:p66=\"b65\" xmlns:p65=\"b64\" xmlns:p64=\"b63\" xmlns:p63=\"b62\" xmlns:p62=\"b61\" xmlns:p61=\"b60\" xmlns:p60=\"b59\" xmlns:p59=\"b58\" xmlns:p58=\"b57\" xmlns:p57=\"b56\" xmlns:p56=\"b55\" xmlns:p55=\"b54\" xmlns:p54=\"b53\" xmlns:p53=\"b52\" xmlns:p52=\"b51\" xmlns:p51=\"b50\" xmlns:p50=\"b49\" xmlns:p49=\"b48\" xmlns:p48=\"b47\" xmlns:p47=\"b46\" xmlns:p46=\"b45\" xmlns:p45=\"b44\" xmlns:p44=\"b43\" xmlns:p43=\"b42\" xmlns:p42=\"b41\" xmlns:p41=\"b40\" xmlns:p40=\"b39\" xmlns:p39=\"b38\" xmlns:p38=\"b37\" xmlns:p37=\"b36\" xmlns:p36=\"b35\" xmlns:p35=\"b34\" xmlns:p34=\"b33\" xmlns:p33=\"b32\" xmlns:p32=\"b31\" xmlns:p31=\"b30\" xmlns:p30=\"b29\" xmlns:p29=\"b28\" xmlns:p28=\"b27\" xmlns:p27=\"b26\" xmlns:p26=\"b25\" xmlns:p25=\"b24\" xmlns:p24=\"b23\" xmlns:p23=\"b22\" xmlns:p22=\"b21\" xmlns:p21=\"b20\" xmlns:p20=\"b19\" xmlns:p19=\"b18\" xmlns:p18=\"b17\" xmlns:p17=\"b16\" xmlns:p16=\"b15\" xmlns:p15=\"b14\" xmlns:p14=\"b13\" xmlns:p13=\"b12\" xmlns:p12=\"b11\" xmlns:p11=\"b10\" xmlns:p10=\"b9\" xmlns:p9=\"b8\" xmlns:p8=\"b7\" xmlns:p7=\"b6\" xmlns:p6=\"b5\" xmlns:p5=\"b4\" xmlns:p4=\"b3\" xmlns:p3=\"b2\" xmlns:p2=\"b1\" xmlns:p1=\"b0\" />";
-            XmlReader r = ReaderHelper.Create(new StringReader(xml));
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(r, (bool)CurVariation.Param);
-            }
-
-            Assert.True((CompareString(xml)));
-        }
-
-        //[Variation(Desc = "Write many attributes and dup namespace")]
-        [Fact]
-        public void element_10a()
-        {
-            XmlWriter w = CreateWriter();
-            w.WriteDocType("a", null, null, "<!ATTLIST oot a CDATA #IMPLIED>");
-            w.WriteStartElement("Root");
-            for (int i = 0; i < 200; i++)
-            {
-                w.WriteAttributeString("a", "n" + i, "val");
-            }
-            try
-            {
-                w.WriteAttributeString("a", "n" + 199, "val");
-            }
-            catch (XmlException) { return; }
-            finally
-            {
-                w.Dispose();
-            }
-            Assert.True(false);
-        }
-
-
-        //[Variation(Desc = "Write many attributes and dup name")]
-        [Fact]
-        public void element_10b()
-        {
-            XmlWriter w = CreateWriter();
-            w.WriteDocType("a", null, null, "<!ATTLIST Root a CDATA #FIXED \"val\">");
-            w.WriteStartElement("Root");
-            for (int i = 0; i < 200; i++)
-            {
-                w.WriteAttributeString("a" + i, "val");
-            }
-            try
-            {
-                w.WriteAttributeString("a" + 199, "val");
-            }
-            catch (XmlException) { return; }
-            finally
-            {
-                w.Dispose();
-            }
-            Assert.True(false);
-        }
-
-        //[Variation(Desc = "Write many attributes and dup prefix")]
-        [Fact]
-        public void element_10c()
-        {
-            XmlWriter w = CreateWriter();
-            w.WriteDocType("a", null, null, "<!ATTLIST Root a (val|value) \"val\">");
-            w.WriteStartElement("Root");
-            for (int i = 0; i < 200; i++)
-            {
-                w.WriteAttributeString("p", "a", "n" + i, "val");
-            }
-            try
-            {
-                w.WriteAttributeString("p", "a", "n" + 199, "val");
-            }
-            catch (XmlException) { return; }
-            finally
-            {
-                w.Dispose();
-            }
-            Assert.True(false);
-        }
-
-        //[Variation(Desc = "Write invalid DOCTYPE with many attributes with prefix")]
-        [Fact]
-        public void element_10d()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteDocType("a", "b", "c", "d");
-                    w.WriteStartElement("Root");
-                    for (int i = 0; i < 200; i++)
-                    {
-                        w.WriteAttributeString("p", "a", "n" + i, "val");
-                    }
-                    w.Dispose();
-                }
-                catch (XmlException e)
-                {
-                    CError.WriteLine(e);
-                    Assert.True(false);
-                }
-            }
-            return;
-        }
-
-        //[Variation(Desc = "WriteEntityRef with XmlWellformedWriter for 'apos'", Param = 1)]
-        //[Variation(Desc = "WriteEntityRef with XmlWellformedWriter for 'lt'", Param = 2)]
-        //[Variation(Desc = "WriteEntityRef with XmlWellformedWriter for 'quot'", Param = 3)]
-        [Fact]
-        public void element_11()
-        {
-            string exp = "";
-            int param = (int)CurVariation.Param;
-            bool isIndent = false;
-            if (WriterType == WriterType.UTF8WriterIndent || WriterType == WriterType.UnicodeWriterIndent)
-                isIndent = true;
-
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteDocType("root", null, null, "<!ENTITY e \"en-us\">");
-                w.WriteStartElement("root");
-                w.WriteStartAttribute("xml", "lang", null);
-                switch (param)
-                {
-                    case 1:
-                        w.WriteEntityRef("apos");
-                        exp = !isIndent ? "<!DOCTYPE root [<!ENTITY e \"en-us\">]><root xml:lang=\"&apos;&lt;\" />" :
-                            "<!DOCTYPE root [<!ENTITY e \"en-us\">]>" + nl + "<root xml:lang=\"&apos;&lt;\" />";
-                        break;
-                    case 2:
-                        w.WriteEntityRef("lt");
-                        exp = !isIndent ? "<!DOCTYPE root [<!ENTITY e \"en-us\">]><root xml:lang=\"&lt;&lt;\" />" :
-                            "<!DOCTYPE root [<!ENTITY e \"en-us\">]>" + nl + "<root xml:lang=\"&lt;&lt;\" />";
-                        break;
-                    case 3:
-                        w.WriteEntityRef("quot");
-                        exp = !isIndent ? "<!DOCTYPE root [<!ENTITY e \"en-us\">]><root xml:lang=\"&quot;&lt;\" />" :
-                            "<!DOCTYPE root [<!ENTITY e \"en-us\">]>" + nl + "<root xml:lang=\"&quot;&lt;\" />";
-                        break;
-                }
-                w.WriteString("<");
-                w.WriteEndAttribute();
-                w.WriteEndElement();
-            }
-            Assert.True((CompareString(exp)));
-        }
-
-        //[Variation(Desc = "WriteValue & WriteWhitespace on a special attribute value – xml:xmlns", Param = 1)]
-        //[Variation(Desc = "WriteValue & WriteWhitespace on a special attribute value – xml:space", Param = 2)]
-        //[Variation(Desc = "WriteValue & WriteWhitespace on a special attribute value – xml:lang", Param = 3)]
-        //[Variation(Desc = "WriteValue & WriteWhitespace on a special attribute value – xmlns", Param = 4)]
-        //[Variation(Desc = "WriteValue & WriteWhitespace on a special attribute value – space", Param = 5)]
-        //[Variation(Desc = "WriteValue & WriteWhitespace on a special attribute value – lang", Param = 6)]
-        [Fact]
-        public void element_12()
-        {
-            int param = (int)CurVariation.Param;
-            XmlWriter w = CreateWriter();
-            w.WriteStartElement("Root");
-            string exp = "";
-            switch (param)
-            {
-                case 1:
-                    exp = "<Root xml:xmlns=\"default    \" />";
-                    break;
-                case 2:
-                    exp = "<Root xml:space=\"default\" />";
-                    break;
-                case 3:
-                    exp = "<Root xml:lang=\"default    \" />";
-                    break;
-                case 4:
-                    exp = "<Root p1:xml=\"default    \" xmlns:p1=\"xmlns\" />";
-                    break;
-                case 5:
-                    exp = "<Root p1:xml=\"default    \" xmlns:p1=\"space\" />";
-                    break;
-                case 6:
-                    exp = "<Root p1:xml=\"default    \" xmlns:p1=\"lang\" />";
-                    break;
-            }
-
-            switch (param)
-            {
-                case 1:
-                    w.WriteStartAttribute("xml", "xmlns", null);
-                    break;
-                case 2:
-                    w.WriteStartAttribute("xml", "space", null);
-                    break;
-                case 3:
-                    w.WriteStartAttribute("xml", "lang", null);
-                    break;
-                case 4:
-                    w.WriteStartAttribute("xml", "xmlns");
-                    break;
-                case 5:
-                    w.WriteStartAttribute("xml", "space");
-                    break;
-                case 6:
-                    w.WriteStartAttribute("xml", "lang");
-                    break;
-            }
-            w.WriteValue("default");
-            try
-            {
-                w.WriteWhitespace("    ");
-                w.WriteEndAttribute();
-                w.WriteEndElement();
-                w.Dispose();
-            }
-            catch (InvalidOperationException e)
-            {
-                CError.WriteLine(e);
-                Assert.True(false);
-            }
-            Assert.True(CompareString(exp));
-        }
-
-        //[Variation(Desc = "WriteValue element double value", Params = new object[] { false, "<Root>-0</Root>" })]
-        //[Variation(Desc = "WriteValue attribute double value", Params = new object[] { true, "<Root b=\"-0\" />" })]
-        [Fact]
-        public void element_13()
-        {
-            bool isAttr = (bool)CurVariation.Params[0];
-            string exp = (string)CurVariation.Params[1];
-            double a = 1;
-            double b = 0;
-            double c = 1;
-            double d = -a * b / c;
-
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                if (isAttr)
-                    w.WriteStartAttribute("b");
-                w.WriteValue(d);
-                w.WriteEndElement();
-            }
-            Assert.True((CompareString(exp)));
-        }
-
-        //[Variation(Desc = "WriteNode with euc-jp encoding.pr-xml-euc-jp.xml", Param = "pr-xml-euc-jp.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.pr-xml-iso-2022-jp.xml", Param = "pr-xml-iso-2022-jp.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.pr-xml-little-endian.xml", Param = "pr-xml-little-endian.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.pr-xml-shift_jis.xml", Param = "pr-xml-shift_jis.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.pr-xml-utf-8.xml", Param = "pr-xml-utf-8.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.pr-xml-utf-16.xml", Param = "pr-xml-utf-16.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.weekly-euc-jp.xml", Param = "weekly-euc-jp.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.weekly-iso-2022-jp.xml", Param = "weekly-iso-2022-jp.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.weekly-little-endian.xml", Param = "weekly-little-endian.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.weekly-shift_jis.xml", Param = "weekly-shift_jis.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.weekly-utf-8.xml", Param = "weekly-utf-8.xml")]
-        //[Variation(Desc = "WriteNode with euc-jp encoding.weekly-utf-16.xml", Param = "weekly-utf-16.xml")]
-        [Fact]
-        public void element_bug480250()
-        {
-            string path = FilePathUtil.GetStandardPath();
-            string xml = (string)CurVariation.Param;
-            string uri = path + @"\XML10\xmlconf\japanese\" + xml;
-            XmlReaderSettings rs = new XmlReaderSettings();
-            rs.IgnoreWhitespace = true;
-            using (XmlReader r = ReaderHelper.Create(uri, rs))
-            {
-                XmlWriterSettings ws = new XmlWriterSettings();
-                ws.Encoding = System.Text.Encoding.GetEncoding("euc-jp");
-                using (XmlWriter w = WriterHelper.Create(@"out.xml", ws))
-                {
-                    w.WriteNode(r, true);
-                }
-            }
-            return;
-        }
-    }
-
-    //[TestCase(Name = "WriteStart/EndAttribute")]
-    public partial class TCAttribute : XmlWriterTestCaseBase
-    {
-        //[Variation(id = 1, Desc = "Sanity test for WriteAttribute", Pri = 0)]
-        [Fact]
-        public void attribute_1()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteStartAttribute("attr1");
-                w.WriteEndAttribute();
-                w.WriteAttributeString("attr2", "val2");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root attr1=\"\" attr2=\"val2\" />"));
-        }
-
-        //[Variation(id = 2, Desc = "Missing EndAttribute should be fixed", Pri = 0)]
-        [Fact]
-        public void attribute_2()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteStartAttribute("attr1");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root attr1=\"\" />"));
-        }
-
-        //[Variation(id = 3, Desc = "WriteStartAttribute followed by WriteStartAttribute", Pri = 0)]
-        [Fact]
-        public void attribute_3()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteStartAttribute("attr1");
-                w.WriteStartAttribute("attr2");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root attr1=\"\" attr2=\"\" />"));
-        }
-
-        //[Variation(id = 4, Desc = "Multiple WritetAttributeString", Pri = 0)]
-        [Fact]
-        public void attribute_4()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteAttributeString("attr1", "val1");
-                w.WriteAttributeString("attr2", "val2");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root attr1=\"val1\" attr2=\"val2\" />"));
-        }
-
-        //[Variation(id = 5, Desc = "WriteStartAttribute followed by WriteString", Pri = 0)]
-        [Fact]
-        public void attribute_5()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteStartAttribute("attr1");
-                w.WriteString("test");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root attr1=\"test\" />"));
-        }
-
-        //[Variation(id = 6, Desc = "Sanity test for overload WriteStartAttribute(name, ns)", Pri = 1)]
-        [Fact]
-        public void attribute_6()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteStartAttribute("attr1", "http://my.com");
-                w.WriteEndAttribute();
-                w.WriteEndElement();
-            }
-            Assert.True(CompareString("<Root ~a p1 a~:attr1=\"\" xmlns:~a p1 A~=\"http://my.com\" />"));
-        }
-
-        //[Variation(id = 7, Desc = "Sanity test for overload WriteStartAttribute(prefix, name, ns)", Pri = 0)]
-        [Fact]
-        public void attribute_7()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteStartAttribute("pre1", "attr1", "http://my.com");
-                w.WriteEndAttribute();
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root pre1:attr1=\"\" xmlns:pre1=\"http://my.com\" />"));
-        }
-
-        //[Variation(id = 8, Desc = "DCR 64183: Duplicate attribute 'attr1'", Pri = 1)]
-        [Fact]
-        public void attribute_8()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteStartAttribute("attr1");
-                    w.WriteStartAttribute("attr1");
-                }
-                catch (XmlException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 9, Desc = "Duplicate attribute 'ns1:attr1'", Pri = 1)]
-        [Fact]
-        public void attribute_9()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteStartAttribute("ns1", "attr1", "http://my.com");
-                    w.WriteStartAttribute("ns1", "attr1", "http://my.com");
-                }
-                catch (XmlException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 10, Desc = "Attribute name = String.Empty should error", Pri = 1)]
-        [Fact]
-        public void attribute_10()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteStartAttribute(String.Empty);
-                }
-                catch (ArgumentException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, (WriterType == WriterType.CharCheckingWriter) ? WriteState.Element : WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 11, Desc = "Attribute name = null", Pri = 1)]
-        [Fact]
-        public void attribute_11()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteStartAttribute(null);
-                }
-                catch (ArgumentException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, (WriterType == WriterType.CharCheckingWriter) ? WriteState.Element : WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 12, Desc = "WriteAttribute with names Foo, fOo, foO, FOO", Pri = 1)]
-        [Fact]
-        public void attribute_12()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                string[] attrNames = { "Foo", "fOo", "foO", "FOO" };
-                w.WriteStartElement("Root");
-                for (int i = 0; i < attrNames.Length; i++)
-                {
-                    w.WriteAttributeString(attrNames[i], "x");
-                }
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root Foo=\"x\" fOo=\"x\" foO=\"x\" FOO=\"x\" />"));
-        }
-
-        //[Variation(id = 13, Desc = "Invalid value of xml:space", Pri = 1)]
-        [Fact]
-        public void attribute_13()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteAttributeString("xml", "space", "http://www.w3.org/XML/1998/namespace", "invalid");
-                }
-                catch (ArgumentException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 14, Desc = "SingleQuote in attribute value should be allowed")]
-        [Fact]
-        public void attribute_14()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteAttributeString("a", null, "b'c");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root a=\"b'c\" />"));
-        }
-
-        //[Variation(id = 15, Desc = "DoubleQuote in attribute value should be escaped")]
-        [Fact]
-        public void attribute_15()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteAttributeString("a", null, "b\"c");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root a=\"b&quot;c\" />"));
-        }
-
-        //[Variation(id = 16, Desc = "WriteAttribute with value = &, #65, #x20", Pri = 1)]
-        [Fact]
-        public void attribute_16()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteAttributeString("a", "&");
-                w.WriteAttributeString("b", "&#65;");
-                w.WriteAttributeString("c", "&#x43;");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root a=\"&amp;\" b=\"&amp;#65;\" c=\"&amp;#x43;\" />"));
-        }
-
-        //[Variation(id = 17, Desc = "WriteAttributeString followed by WriteString", Pri = 1)]
-        [Fact]
-        public void attribute_17()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteAttributeString("a", null, "b");
-                w.WriteString("test");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root a=\"b\">test</Root>"));
-        }
-
-        //[Variation(id = 18, Desc = "WriteAttribute followed by WriteString", Pri = 1)]
-        [Fact]
-        public void attribute_18()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteStartAttribute("a");
-                w.WriteString("test");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root a=\"test\" />"));
-        }
-
-        //[Variation(id = 19, Desc = "WriteAttribute with all whitespace characters", Pri = 1)]
-        [Fact]
-        public void attribute_19()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteAttributeString("a", null, "\x20\x9\xD\xA");
-                w.WriteEndElement();
-            }
-
-            Assert.True(CompareReader("<Root a=\" &#x9;&#xD;&#xA;\" />"));
-        }
-
-        //[Variation(id = 20, Desc = "< > & chars should be escaped in attribute value", Pri = 1)]
-        [Fact]
-        public void attribute_20()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteAttributeString("a", null, "< > &");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root a=\"&lt; &gt; &amp;\" />"));
-        }
-
-        //[Variation(id = 21, Desc = "Redefine auto generated prefix n1")]
-        [Fact]
-        public void attribute_21()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("test");
-                w.WriteAttributeString("xmlns", "n1", null, "http://testbasens");
-                w.WriteStartElement("base");
-                w.WriteAttributeString("id", "http://testbasens", "5");
-                w.WriteAttributeString("lang", "http://common", "en");
-                w.WriteEndElement();
-                w.WriteEndElement();
-            }
-            string exp = IsIndent() ?
-                "<test xmlns:n1=\"http://testbasens\">" + Environment.NewLine + "  <base n1:id=\"5\" p4:lang=\"en\" xmlns:p4=\"http://common\" />" + Environment.NewLine + "</test>" :
-                "<test xmlns:~f n1 A~=\"http://testbasens\"><base ~f n1 a~:id=\"5\" ~a p4 a~:lang=\"en\" xmlns:~a p4 A~=\"http://common\" /></test>";
-            Assert.True(CompareString(exp));
-        }
-
-        //[Variation(id = 22, Desc = "Reuse and redefine existing prefix")]
-        [Fact]
-        public void attribute_22()
-        {
-            string exp = "<test ~f p a~:a1=\"v\" xmlns:~f p A~=\"ns1\"><base ~f p b~:a2=\"v\" ~a p4 ab~:a3=\"v\" xmlns:~a p4 AB~=\"ns2\" /></test>";
-
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("test");
-                w.WriteAttributeString("p", "a1", "ns1", "v");
-                w.WriteStartElement("base");
-                w.WriteAttributeString("a2", "ns1", "v");
-                w.WriteAttributeString("p", "a3", "ns2", "v");
-                w.WriteEndElement();
-                w.WriteEndElement();
-            }
-            exp = IsIndent() ?
-                "<test p:a1=\"v\" xmlns:p=\"ns1\">" + Environment.NewLine + "  <base p:a2=\"v\" p4:a3=\"v\" xmlns:p4=\"ns2\" />" + Environment.NewLine + "</test>" : exp;
-            Assert.True(CompareString(exp));
-        }
-
-        //[Variation(id = 23, Desc = "WriteStartAttribute(attr) sanity test")]
-        [Fact]
-        public void attribute_23()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("test");
-                w.WriteStartAttribute("attr");
-                w.WriteEndAttribute();
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<test attr=\"\" />"));
-        }
-
-        //[Variation(id = 24, Desc = "WriteStartAttribute(attr) inside an element with changed default namespace")]
-        [Fact]
-        public void attribute_24()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement(string.Empty, "test", "ns");
-                w.WriteStartAttribute("attr");
-                w.WriteEndAttribute();
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<test attr=\"\" xmlns=\"ns\" />"));
-        }
-
-        //[Variation(id = 25, Desc = "WriteStartAttribute(attr) and duplicate attrs")]
-        [Fact]
-        public void attribute_25()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("test");
-                    w.WriteStartAttribute(null, "attr", null);
-                    w.WriteStartAttribute("attr");
-                }
-                catch (XmlException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw error for duplicate attrs");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 26, Desc = "WriteStartAttribute(attr) when element has ns:attr")]
-        [Fact]
-        public void attribute_26()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("pre", "test", "ns");
-                w.WriteStartAttribute(null, "attr", "ns");
-                w.WriteStartAttribute("attr");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<pre:test pre:attr=\"\" attr=\"\" xmlns:pre=\"ns\" />"));
-        }
-
-        //[Variation(id = 27, Desc = "XmlCharCheckingWriter should not normalize newLines in attribute values when NewLinesHandling = Replace")]
-        [Fact]
-        public void attribute_27()
-        {
-            XmlWriterSettings s = new XmlWriterSettings();
-            s.NewLineHandling = NewLineHandling.Replace;
-
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("root");
-                w.WriteAttributeString("a", "|\x0D|\x0A|\x0D\x0A|");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<root a=\"|&#xD;|&#xA;|&#xD;&#xA;|\" />"));
-        }
-
-        //[Variation(id = 28, Desc = "Wrapped XmlTextWriter: Invalid replacement of newline characters in text values")]
-        [Fact]
-        public void attribute_28()
-        {
-            XmlWriterSettings s = new XmlWriterSettings();
-            s.NewLineHandling = NewLineHandling.Replace;
-
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("root");
-                w.WriteAttributeString("a", "|\x0D\x0A|");
-                w.WriteElementString("foo", "|\x0D\x0A|");
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<root a=\"|&#xD;&#xA;|\"><foo>|\x0D\x0A|</foo></root>"));
-        }
-
-        //[Variation(id = 29, Desc = "WriteAttributeString doesn't fail on invalid surrogate pair sequences")]
-        [Fact]
-        public void attribute_29()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("root");
-                try
-                {
-                    w.WriteAttributeString("attribute", "\ud800\ud800");
-                }
-                catch (ArgumentException e)
-                {
-                    CError.WriteLineIgnore(e.ToString());
-                    return;
-                }
-            }
-            Assert.True(false);
-        }
-    }
-
-    //[TestCase(Name = "WriteAttributes(XmlTextReader)", Param = "XMLTEXTREADER")]
-    //[TestCase(Name = "WriteAttributes(XmlDocument NodeReader)", Param = "XMLDOCNODEREADER")]
-    //[TestCase(Name = "WriteAttributes(XmlDataDocument NodeReader)", Param = "XMLDATADOCNODEREADER")]
-    //[TestCase(Name = "WriteAttributes(XsltReader)", Param = "XSLTREADER")]
-    //[TestCase(Name = "WriteAttributes(CoreReader)", Param = "COREREADER")]
-    //[TestCase(Name = "WriteAttributes(XPathdocument NavigatorReader)", Param = "XPATHDOCNAVIGATORREADER")]
-    //[TestCase(Name = "WriteAttributes(XmlDocument NavigatorReader)", Param = "XMLDOCNAVIGATORREADER")]
-    //[TestCase(Name = "WriteAttributes(XmlDataDocument NavigatorReader)", Param = "XMLDATADOCNAVIGATORREADER")]
-    //[TestCase(Name = "WriteAttributes(XmlBinaryReader)", Param = "XMLBINARYREADER")]
-    public partial class TCWriteAttributes : ReaderParamTestCase
-    {
-        //[Variation(id = 1, Desc = "Call WriteAttributes with default DTD attributes = true", Pri = 1)]
-        [Fact]
-        public void writeAttributes_1()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReader("XmlReader.xml"))
-                {
-                    while (xr.Read())
-                    {
-                        if (xr.LocalName == "name")
-                        {
-                            xr.MoveToFirstAttribute();
-                            break;
-                        }
-                    }
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, true);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareReader("<Root a=\"b\" FIRST=\"KEVIN\" LAST=\"WHITE\" />"));
-        }
-
-        //[Variation(id = 2, Desc = "Call WriteAttributes with default DTD attributes = false", Pri = 1)]
-        [Fact]
-        public void writeAttributes_2()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReader("XmlReader.xml"))
-                {
-                    while (xr.Read())
-                    {
-                        if (xr.LocalName == "name")
-                        {
-                            xr.MoveToFirstAttribute();
-                            break;
-                        }
-                    }
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-
-            if (IsXPathDataModelReader())
-                // always sees default attributes
-                Assert.True(CompareReader("<Root a=\"b\" FIRST=\"KEVIN\" LAST=\"WHITE\" />"));
-            else
-                Assert.True(CompareReader("<Root a=\"b\" />"));
-        }
-
-        //[Variation(id = 3, Desc = "Call WriteAttributes with XmlReader = null")]
-        [Fact]
-        public void writeAttributes_3()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                XmlReader xr = null;
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, false);
-                }
-                catch (ArgumentNullException e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Element, "WriteState should be Element");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 4, Desc = "Call WriteAttributes when reader is located on element", Pri = 1)]
-        [Fact]
-        public void writeAttributes_4()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReader("XmlReader.xml"))
-                {
-                    while (xr.Read())
-                    {
-                        if (xr.LocalName == "AttributesGeneric")
-                        {
-                            do
-                            { xr.Read(); } while (xr.LocalName != "node");
-                            break;
-                        }
-                    }
-                    if (xr.NodeType != XmlNodeType.Element)
-                    {
-                        CError.WriteLine("Reader not positioned element");
-                        CError.WriteLine(xr.LocalName);
-                        xr.Dispose();
-                        w.Dispose();
-                        Assert.True(false);
-                    }
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareReader("<Root a=\"b\" c=\"d\" e=\"f\" />"));
-        }
-
-        //[Variation(id = 5, Desc = "Call WriteAttributes when reader is located in the middle attribute", Pri = 1)]
-        [Fact]
-        public void writeAttributes_5()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReader("XmlReader.xml"))
-                {
-                    while (xr.Read())
-                    {
-                        if (xr.LocalName == "AttributesGeneric")
-                        {
-                            do
-                            { xr.Read(); } while (xr.LocalName != "node");
-                            xr.MoveToAttribute(1);
-                            break;
-                        }
-                    }
-                    if (xr.NodeType != XmlNodeType.Attribute)
-                    {
-                        CError.WriteLine("Reader not positioned on attribute");
-                        CError.WriteLine(xr.LocalName);
-                        xr.Dispose();
-                        w.Dispose();
-                        Assert.True(false);
-                    }
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareReader("<Root c=\"d\" e=\"f\" />"));
-        }
-
-        //[Variation(id = 6, Desc = "Call WriteAttributes when reader is located in the last attribute", Pri = 1)]
-        [Fact]
-        public void writeAttributes_6()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReader("XmlReader.xml"))
-                {
-                    while (xr.Read())
-                    {
-                        if (xr.LocalName == "AttributesGeneric")
-                        {
-                            do
-                            { xr.Read(); } while (xr.LocalName != "node");
-                            xr.MoveToNextAttribute();
-                            xr.MoveToNextAttribute();
-                            xr.MoveToNextAttribute();
-                            break;
-                        }
-                    }
-                    if (xr.NodeType != XmlNodeType.Attribute)
-                    {
-                        CError.WriteLine("Reader not positioned on attribute");
-                        CError.WriteLine(xr.LocalName);
-                        xr.Dispose();
-                        w.Dispose();
-                        Assert.True(false);
-                    }
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareReader("<Root e=\"f\" />"));
-        }
-
-        //[Variation(id = 7, Desc = "Call WriteAttributes when reader is located on an attribute with an entity reference in the value", Pri = 1)]
-        [Fact]
-        public void writeAttributes_7()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReader("XmlReader.xml"))
-                {
-                    while (xr.Read())
-                    {
-                        if (xr.LocalName == "AttributesEntity")
-                        {
-                            do
-                            { xr.Read(); } while (xr.LocalName != "node");
-                            xr.MoveToNextAttribute();
-                            break;
-                        }
-                    }
-                    if (xr.NodeType != XmlNodeType.Attribute)
-                    {
-                        CError.WriteLine("Reader not positioned on attribute");
-                        CError.WriteLine(xr.LocalName);
-                        xr.Dispose();
-                        w.Dispose();
-                        Assert.True(false);
-                    }
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            if (!ReaderExpandsEntityRef())
-                Assert.True(CompareString("<Root a=\"&e;\" />"));
-            else
-                Assert.True(CompareReader("<Root a=\"Test Entity\" />"));
-        }
-
-        //[Variation(id = 8, Desc = "Call WriteAttributes with reader on XmlDeclaration", Pri = 1)]
-        [Fact]
-        public void writeAttributes_8()
-        {
-            if (IsXPathDataModelReader())
-            {
-                CError.WriteLine("{0} does not support XmlDecl node", readerType);
-                return;
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReader("Simple.xml"))
-                {
-                    xr.Read();
-                    if (xr.NodeType != XmlNodeType.XmlDeclaration)
-                    {
-                        CError.WriteLine("Reader not positioned on XmlDeclaration");
-                        CError.WriteLine(xr.LocalName);
-                        xr.Dispose();
-                        w.Dispose();
-                        Assert.True(false);
-                    }
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareReader("<Root version=\"1.0\" standalone=\"yes\" />"));
-        }
-
-        //[Variation(id = 9, Desc = "Call WriteAttributes with reader on DocType", Pri = 1, Param = "DocumentType")]
-        //[Variation(id = 10, Desc = "Call WriteAttributes with reader on CDATA", Pri = 1, Param = "CDATA")]
-        //[Variation(id = 11, Desc = "Call WriteAttributes with reader on Text", Pri = 1, Param = "Text")]
-        //[Variation(id = 12, Desc = "Call WriteAttributes with reader on PI", Pri = 1, Param = "ProcessingInstruction")]
-        //[Variation(id = 13, Desc = "Call WriteAttributes with reader on Comment", Pri = 1, Param = "Comment")]
-        //[Variation(id = 14, Desc = "Call WriteAttributes with reader on EntityRef", Pri = 1, Param = "EntityReference")]
-        //[Variation(id = 15, Desc = "Call WriteAttributes with reader on Whitespace", Pri = 1, Param = "Whitespace")]
-        //[Variation(id = 16, Desc = "Call WriteAttributes with reader on SignificantWhitespace", Pri = 1, Param = "SignificantWhitespace")]
-        [Fact]
-        public void writeAttributes_9()
-        {
-            string strxml = "";
-            switch (CurVariation.Param.ToString())
-            {
-                case "DocumentType":
-                    if (IsXPathDataModelReader())
-                    {
-                        CError.WriteLine("{0} does not support DocumentType node", readerType);
-                        return;
-                    }
-                    strxml = "<!DOCTYPE Root[]><Root/>";
-                    break;
-                case "CDATA":
-                    if (IsXPathDataModelReader())
-                    {
-                        CError.WriteLine("{0} does not support CDATA node", readerType);
-                        return;
-                    }
-                    strxml = "<root><![CDATA[Test]]></root>";
-                    break;
-                case "Text":
-                    strxml = "<root>Test</root>";
-                    break;
-                case "ProcessingInstruction":
-                    strxml = "<root><?pi test?></root>";
-                    break;
-                case "Comment":
-                    strxml = "<root><!-- comment --></root>";
-                    break;
-                case "EntityReference":
-                    if (!ReaderSupportsEntityRef())
-                    {
-                        CError.WriteLine("{0} does not support EntityRef node", readerType);
-                        return;
-                    }
-                    strxml = "<!DOCTYPE root[<!ENTITY e \"Test Entity\"> ]><root>&e;</root>";
-                    break;
-                case "SignificantWhitespace":
-                    strxml = "<root xml:space=\"preserve\">			 </root>";
-                    break;
-                case "Whitespace":
-                    if (ReaderStripsWhitespace())
-                    {
-                        CError.WriteLine("{0} strips whitespace nodes by default", readerType);
-                        return;
-                    }
-                    strxml = "<root>			 </root>";
-                    break;
-            }
-
-            XmlReader xr;
-            xr = CreateReader(new StringReader(strxml));
-
-            do
-            { xr.Read(); }
-            while ((xr.NodeType.ToString() != CurVariation.Param.ToString()) && (xr.ReadState != ReadState.EndOfFile));
-
-            if (xr.ReadState == ReadState.EndOfFile || xr.NodeType.ToString() != CurVariation.Param.ToString())
-            {
-                xr.Dispose();
-                CError.WriteLine("Reader not positioned on correct node");
-                CError.WriteLine("ReadState: {0}", xr.ReadState);
-                CError.WriteLine("NodeType: {0}", xr.NodeType);
-                Assert.True(false);
-            }
-
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    if (CurVariation.Param.ToString() != "DocumentType")
-                        w.WriteStartElement("root");
-                    w.WriteAttributes(xr, false);
-                }
-                catch (XmlException e)
-                {
-                    CError.WriteLineIgnore(e.ToString());
-                    CError.Compare(w.WriteState, (CurVariation.Param.ToString() == "DocumentType") ? WriteState.Start : WriteState.Element, "WriteState should be Element");
-                    return;
-                }
-                finally
-                {
-                    xr.Dispose();
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 17, Desc = "Call WriteAttribute with double quote char in the value", Pri = 1)]
-        [Fact]
-        public void writeAttributes_10()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReader("XmlReader.xml"))
-                {
-                    while (xr.Read())
-                    {
-                        if (xr.LocalName == "QuoteChar")
-                        {
-                            do
-                            { xr.Read(); } while (xr.LocalName != "node");
-                            xr.MoveToFirstAttribute();
-                            break;
-                        }
-                    }
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareReader("<Root a=\"b&quot;c\" />"));
-        }
-
-        //[Variation(id = 18, Desc = "Call WriteAttribute with single quote char in the value", Pri = 1)]
-        [Fact]
-        public void writeAttributes_11()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReader("XmlReader.xml"))
-                {
-                    while (xr.Read())
-                    {
-                        if (xr.LocalName == "QuoteChar")
-                        {
-                            do
-                            { xr.Read(); } while (xr.LocalName != "node");
-                            do
-                            { xr.Read(); } while (xr.LocalName != "node");
-                            xr.MoveToFirstAttribute();
-                            break;
-                        }
-                    }
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareReader("<Root a=\"b'c\" />"));
-        }
-
-        //[Variation(id = 19, Desc = "Call WriteAttributes with 100 attributes", Pri = 1)]
-        [Fact]
-        public void writeAttributes_12()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReader("XmlReader.xml"))
-                {
-                    while (xr.Read())
-                    {
-                        if (xr.LocalName == "OneHundredAttributes")
-                        {
-                            xr.MoveToFirstAttribute();
-                            break;
-                        }
-                    }
-
-                    if (xr.NodeType != XmlNodeType.Attribute)
-                    {
-                        CError.WriteLine("Reader positioned on {0}", xr.NodeType.ToString());
-                        xr.Dispose();
-                        Assert.True(false);
-                    }
-                    w.WriteStartElement("OneHundredAttributes");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareBaseline("OneHundredAttributes.xml"));
-        }
-
-        //[Variation(id = 20, Desc = "WriteAttributes with different builtin entities in attribute value", Pri = 1)]
-        [Fact]
-        public void writeAttributes_13()
-        {
-            string strxml = "<E a=\"&gt;&lt;&quot;&apos;&amp;\" />";
-            using (XmlReader xr = CreateReader(new StringReader(strxml)))
-            {
-                xr.Read();
-                xr.MoveToFirstAttribute();
-
-                if (xr.NodeType != XmlNodeType.Attribute)
-                {
-                    CError.WriteLine("Reader positioned on {0}", xr.NodeType.ToString());
-                    xr.Dispose();
-                    Assert.True(false);
-                }
-
-                using (XmlWriter w = CreateWriter())
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareReader("<Root a=\"&gt;&lt;&quot;&apos;&amp;\" />"));
-        }
-
-        //[Variation(id = 21, Desc = "WriteAttributes tries to duplicate attribute", Pri = 1)]
-        [Fact]
-        public void writeAttributes_14()
-        {
-            string strxml = "<root attr='test' />";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            xr.Read();
-            xr.MoveToFirstAttribute();
-
-            if (xr.NodeType != XmlNodeType.Attribute)
-            {
-                CError.WriteLine("Reader positioned on {0}", xr.NodeType.ToString());
-                xr.Dispose();
-                Assert.True(false);
-            }
-
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteStartAttribute("attr");
-                    w.WriteAttributes(xr, false);
-                    w.WriteEndElement();
-                }
-                catch (Exception e)
-                {
-                    CError.WriteLineIgnore("Exception: " + e.ToString());
-                    CError.Compare(w.WriteState, WriteState.Error, "WriteState should be Error");
-                    return;
-                }
-                finally
-                {
-                    xr.Dispose();
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-    }
-
-    //[TestCase(Name = "WriteNode(XmlTextReader)", Param = "XMLTEXTREADER")]
-    //[TestCase(Name = "WriteNode(XmlDocument NodeReader)", Param = "XMLDOCNODEREADER")]
-    //[TestCase(Name = "WriteNode(XmlDataDocument NodeReader)", Param = "XMLDATADOCNODEREADER")]
-    //[TestCase(Name = "WriteNode(XsltReader)", Param = "XSLTREADER")]
-    //[TestCase(Name = "WriteNode(CoreReader)", Param = "COREREADER")]
-    //[TestCase(Name = "WriteNode(XPathdocument NavigatorReader)", Param = "XPATHDOCNAVIGATORREADER")]
-    //[TestCase(Name = "WriteNode(XmlDocument NavigatorReader)", Param = "XMLDOCNAVIGATORREADER")]
-    //[TestCase(Name = "WriteNode(XmlDataDocument NavigatorReader)", Param = "XMLDATADOCNAVIGATORREADER")]
-    //[TestCase(Name = "WriteNode(XmlBinaryReader)", Param = "XMLBINARYREADER")]
-    public partial class TCWriteNode_XmlReader : ReaderParamTestCase
-    {
-        //[Variation(id = 1, Desc = "WriteNode with null reader", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader1()
-        {
-            XmlReader xr = null;
-            using (XmlWriter w = CreateWriter())
-            {
-                try
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteNode(xr, false);
-                }
-                catch (ArgumentNullException)
-                {
-                    CError.Compare(w.WriteState, WriteState.Element, "WriteState should be Element");
-                    return;
-                }
-            }
-            CError.WriteLine("Did not throw exception");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 2, Desc = "WriteNode with reader positioned on attribute, no operation", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader2()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                using (XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml"))
-                {
-                    while (xr.Read())
-                    {
-                        if (xr.LocalName == "defattr")
-                        {
-                            xr.Read();
-                            xr.MoveToFirstAttribute();
-                            break;
-                        }
-                    }
-
-                    if (xr.NodeType != XmlNodeType.Attribute)
-                    {
-                        CError.WriteLine("Reader positioned on {0}", xr.NodeType.ToString());
-                        xr.Dispose();
-                        w.Dispose();
-                        Assert.True(false);
-                    }
-                    w.WriteStartElement("Root");
-                    w.WriteNode(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareReader("<Root />"));
-        }
-
-        //[Variation(id = 3, Desc = "WriteNode before reader.Read()", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader3()
-        {
-            using (XmlReader xr = CreateReader(new StringReader("<root />")))
-            {
-                using (XmlWriter w = CreateWriter())
-                {
-                    w.WriteNode(xr, false);
-                }
-            }
-
-            Assert.True(CompareReader("<root />"));
-        }
-
-        //[Variation(id = 4, Desc = "WriteNode after first reader.Read()", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader4()
-        {
-            using (XmlReader xr = CreateReader(new StringReader("<root />")))
-            {
-                xr.Read();
-                using (XmlWriter w = CreateWriter())
-                {
-                    w.WriteNode(xr, false);
-                }
-            }
-
-            Assert.True(CompareReader("<root />"));
-        }
-
-        //[Variation(id = 5, Desc = "WriteNode when reader is positioned on middle of an element node", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader5()
-        {
-            using (XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml"))
-            {
-                while (xr.Read())
-                {
-                    if (xr.LocalName == "Middle")
-                    {
-                        xr.Read();
-                        break;
-                    }
-                }
-                using (XmlWriter w = CreateWriter())
-                {
-                    w.WriteNode(xr, false);
-                }
-                CError.Compare(xr.NodeType, XmlNodeType.Comment, "Error");
-                CError.Compare(xr.Value, "WriteComment", "Error");
-            }
-            Assert.True(CompareReader("<node2>Node Text<node3></node3><?name Instruction?></node2>"));
-        }
-
-        //[Variation(id = 6, Desc = "WriteNode when reader state is EOF", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader6()
-        {
-            using (XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml"))
-            {
-                while (xr.Read())
-                { }
-
-                using (XmlWriter w = CreateWriter())
-                {
-                    w.WriteStartElement("Root");
-                    w.WriteNode(xr, false);
-                    w.WriteEndElement();
-                }
-            }
-            Assert.True(CompareReader("<Root />"));
-        }
-
-        //[Variation(id = 7, Desc = "WriteNode when reader state is Closed", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader7()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            { }
-            xr.Dispose();
-
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteNode(xr, false);
-                w.WriteEndElement();
-            }
-            Assert.True(CompareReader("<Root />"));
-        }
-
-        //[Variation(id = 8, Desc = "WriteNode with reader on empty element node", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader8()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "EmptyElement")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-            }
-
-            // check reader position
-            CError.Compare(xr.NodeType, XmlNodeType.EndElement, "Error");
-            CError.Compare(xr.Name, "EmptyElement", "Error");
-            xr.Dispose();
-
-            Assert.True(CompareReader("<node1 />"));
-        }
-
-        //[Variation(id = 9, Desc = "WriteNode with reader on 100 Nodes", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader9()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "OneHundredElements")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-            }
-            xr.Dispose();
-            Assert.True(CompareBaseline("100Nodes.txt"));
-        }
-
-        //[Variation(id = 10, Desc = "WriteNode with reader on node with mixed content", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader10()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "MixedContent")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-            }
-
-            // check reader position
-            CError.Compare(xr.NodeType, XmlNodeType.EndElement, "Error");
-            CError.Compare(xr.Name, "MixedContent", "Error");
-            xr.Dispose();
-
-            if (IsXPathDataModelReader())
-            {
-                Assert.True(CompareReader("<node1><?PI Instruction?><!--Comment-->Textcdata</node1>"));
-            }
-
-            Assert.True(CompareReader("<node1><?PI Instruction?><!--Comment-->Text<![CDATA[cdata]]></node1>"));
-        }
-
-        //[Variation(id = 11, Desc = "WriteNode with reader on node with declared namespace in parent", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader11()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "NamespaceNoPrefix")
-                {
-                    xr.Read();
-                    xr.Read();
-                    break;
-                }
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-            }
-            xr.Dispose();
-            Assert.True(CompareReader("<node1 xmlns=\"foo\"></node1>"));
-        }
-
-        //[Variation(id = 12, Desc = "WriteNode with reader on node with entity reference included in element", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader12()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "EntityRef")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-            }
-            xr.Dispose();
-            if (!ReaderExpandsEntityRef())
-                Assert.True(CompareString("<node>&e;</node>"));
-            else
-                Assert.True(CompareReader("<node>Test Entity</node>"));
-        }
-
-        //[Variation(id = 14, Desc = "WriteNode with element that has different prefix", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader14()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "DiffPrefix")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("x", "bar", "foo");
-                w.WriteNode(xr, true);
-                w.WriteStartElement("blah", "foo");
-                w.WriteEndElement();
-                w.WriteEndElement();
-            }
-            xr.Dispose();
-
-            Assert.True(CompareReader("<x:bar xmlns:x=\"foo\"><z:node xmlns:z=\"foo\" /><x:blah /></x:bar>"));
-        }
-
-        //[Variation(id = 15, Desc = "Call WriteNode with default attributes = true and DTD", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader15()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "DefaultAttributesTrue")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteNode(xr, true);
-                w.WriteEndElement();
-            }
-            xr.Dispose();
-            if (!ReaderParsesDTD())
-                Assert.True(CompareReader("<Root><name a='b' /></Root>"));
-            else
-                Assert.True(CompareReader("<Root><name a='b' FIRST='KEVIN' LAST='WHITE'/></Root>"));
-        }
-
-        //[Variation(id = 16, Desc = "Call WriteNode with default attributes = false and DTD", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader16()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "DefaultAttributesTrue")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("Root");
-                w.WriteNode(xr, false);
-                w.WriteEndElement();
-            }
-            xr.Dispose();
-            if (ReaderLoosesDefaultAttrInfo())
-                Assert.True(CompareReader("<Root><name a='b' FIRST='KEVIN' LAST='WHITE'/></Root>"));
-            else
-                Assert.True(CompareReader("<Root><name a='b' /></Root>"));
-        }
-
-        //[Variation(id = 17, Desc = "WriteNode with reader on empty element with attributes", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader17()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "EmptyElementWithAttributes")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-            }
-            Assert.True(CompareReader("<node1 a='foo' />"));
-        }
-
-        //[Variation(id = 18, Desc = "WriteNode with document containing just empty element with attributes", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader18()
-        {
-            string xml = "<Root a=\"foo\"/>";
-            XmlReader xr = CreateReader(new StringReader(xml));
-            xr.Read();
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-            }
-
-            xr.Dispose();
-            Assert.True(CompareReader("<Root a=\"foo\" />"));
-        }
-
-        //[Variation(id = 19, Desc = "Call WriteNode with special entity references as attribute value", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader19()
-        {
-            using (XmlWriter w = CreateWriter())
-            {
-                string xml = "<Root foo='&amp; &lt; &gt; &quot; &apos; &#65;'/>";
-                using (XmlReader xr = CreateReader(new StringReader(xml)))
-                {
-                    while (xr.Read())
-                        w.WriteNode(xr, true);
-                }
-            }
-            Assert.True(CompareReader("<Root foo='&amp; &lt; &gt; &quot; &apos; &#65;'/>"));
-        }
-
-        //[Variation(id = 20, Desc = "Call WriteNode with reader on doctype", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader20()
-        {
-            string strxml = "<!DOCTYPE ROOT []><ROOT/>";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            using (XmlWriter w = CreateWriter())
-            {
-                while (xr.NodeType != XmlNodeType.DocumentType)
-                    xr.Read();
-
-                w.WriteNode(xr, false);
-                w.WriteStartElement("ROOT");
-                w.WriteEndElement();
-            }
-            xr.Dispose();
-
-            Assert.True(CompareReader("<!DOCTYPE ROOT[]><ROOT />"));
-        }
-
-        //[Variation(id = 21, Desc = "Call WriteNode with full end element", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader21()
-        {
-            string strxml = "<root></root>";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-            }
-
-            xr.Dispose();
-            Assert.True(CompareReader("<root></root>"));
-        }
-
-        //[Variation(Desc = "Call WriteNode with tag mismatch")]
-        [Fact]
-        public void writeNode_XmlReader21a()
-        {
-            string strxml = "<a xmlns=\"p1\"><b xmlns=\"p2\"><c xmlns=\"p1\" /></b><d xmlns=\"\"><e xmlns=\"p1\"><f xmlns=\"\" /></d></a>";
-            try
-            {
-                using (XmlWriter w = CreateWriter())
-                {
-                    try
-                    {
-                        using (XmlReader xr = CreateReader(new StringReader(strxml)))
-                        {
-                            w.WriteNode(xr, true);
-                            CError.Compare(false, "Failed");
-                        }
-                    }
-                    catch (XmlException xe) { CError.WriteLine(xe.Message); return; }
-                }
-            }
-            catch (ObjectDisposedException e) { CError.WriteLine(e.Message); return; }
-            Assert.True(false);
-        }
-
-        //[Variation(Desc = "Call WriteNode with default NS from DTD.UnexpToken")]
-        [Fact]
-        public void writeNode_XmlReader21b()
-        {
-            string strxml = "<!DOCTYPE doc " +
-"[<!ELEMENT doc ANY>" +
-"<!ELEMENT test1 (#PCDATA)>" +
-"<!ELEMENT test2 ANY>" +
-"<!ELEMENT test3 (#PCDATA)>" +
-"<!ENTITY e1 \"&e2;\">" +
-"<!ENTITY e2 \"xmlns=\"x\"\">" +
-"<!ATTLIST test3 a1 CDATA #IMPLIED>" +
-"<!ATTLIST test3 a2 CDATA #IMPLIED>" +
-"]>" +
-"<doc>" +
-"    &e2;" +
-"    <test1>AA&e2;AA</test1>" +
-"    <test2>BB&e1;BB</test2>" +
-"    <test3 a1=\"&e2;\" a2=\"&e1;\">World</test3>" +
-"</doc>";
-            try
-            {
-                using (XmlWriter w = CreateWriter())
-                {
-                    try
-                    {
-                        using (XmlReader xr = CreateReader(new StringReader(strxml)))
-                        {
-                            w.WriteNode(xr, true);
-                            CError.Compare(false, "Failed");
-                        }
-                    }
-                    catch (XmlException xe) { CError.WriteLine(xe.Message); return; }
-                }
-            }
-            catch (ObjectDisposedException e) { CError.WriteLine(e.Message); return; }
-            Assert.True(false);
-        }
-
-        //[Variation(id = 22, Desc = "Call WriteNode with reader on element with 100 attributes", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader22()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "OneHundredAttributes")
-                {
-                    break;
-                }
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-            }
-            xr.Dispose();
-            Assert.True(CompareBaseline("OneHundredAttributes.xml"));
-        }
-
-        //[Variation(id = 23, Desc = "Call WriteNode with reader on text node", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader23()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "Middle")
-                {
-                    xr.Read();
-                    xr.Read();
-                    break;
-                }
-            }
-            if (xr.NodeType != XmlNodeType.Text)
-            {
-                CError.WriteLine("Reader positioned on {0}", xr.NodeType);
-                xr.Dispose();
-                Assert.True(false);
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("root");
-                w.WriteNode(xr, false);
-                w.WriteEndElement();
-            }
-            xr.Dispose();
-            Assert.True(CompareReader("<root>Node Text</root>"));
-        }
-
-        //[Variation(id = 24, Desc = "Call WriteNode with reader on CDATA node", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader24()
-        {
-            if (IsXPathDataModelReader())
-            {
-                CError.WriteLine("XPath data model does not have CDATA node type, so {0} can not be positioned on CDATA", readerType);
-                return;
-            }
-
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "CDataNode")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            if (xr.NodeType != XmlNodeType.CDATA)
-            {
-                CError.WriteLine("Reader positioned on {0}", xr.NodeType);
-                xr.Dispose();
-                Assert.True(false);
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("root");
-                w.WriteNode(xr, false);
-                w.WriteEndElement();
-            }
-            xr.Dispose();
-
-            Assert.True(CompareReader("<root><![CDATA[cdata content]]></root>"));
-        }
-
-        //[Variation(id = 25, Desc = "Call WriteNode with reader on PI node", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader25()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "PINode")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            if (xr.NodeType != XmlNodeType.ProcessingInstruction)
-            {
-                CError.WriteLine("Reader positioned on {0}", xr.NodeType);
-                xr.Dispose();
-                Assert.True(false);
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("root");
-                w.WriteNode(xr, false);
-                w.WriteEndElement();
-            }
-            xr.Dispose();
-            Assert.True(CompareReader("<root><?PI Text?></root>"));
-        }
-
-        //[Variation(id = 26, Desc = "Call WriteNode with reader on Comment node", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader26()
-        {
-            XmlReader xr = CreateReaderIgnoreWS("XmlReader.xml");
-            while (xr.Read())
-            {
-                if (xr.LocalName == "CommentNode")
-                {
-                    xr.Read();
-                    break;
-                }
-            }
-            if (xr.NodeType != XmlNodeType.Comment)
-            {
-                CError.WriteLine("Reader positioned on {0}", xr.NodeType);
-                xr.Dispose();
-                Assert.True(false);
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteStartElement("root");
-                w.WriteNode(xr, false);
-                w.WriteEndElement();
-            }
-            xr.Dispose();
-            Assert.True(CompareReader("<root><!--Comment--></root>"));
-        }
-
-        //[Variation(Desc = "Call WriteNode with reader on XmlDecl (OmitXmlDecl false)", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader28()
-        {
-            string strxml = "<?xml version=\"1.0\" standalone=\"yes\"?><Root />";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-
-            xr.Read();
-            if (xr.NodeType != XmlNodeType.XmlDeclaration)
-            {
-                CError.WriteLine("Reader positioned on {0}", xr.NodeType);
-                xr.Dispose();
-                return;
-            }
-
-            XmlWriterSettings ws = new XmlWriterSettings();
-            ws.OmitXmlDeclaration = false;
-            XmlWriter w = CreateWriter(ws);
-            w.WriteNode(xr, false);
-            w.WriteStartElement("Root");
-            w.WriteEndElement();
-            xr.Dispose();
-            w.Dispose();
-            strxml = IsIndent() ? "<?xml version=\"1.0\" standalone=\"yes\"?>" + Environment.NewLine + "<Root />" : strxml;
-            Assert.True(CompareString(strxml));
-        }
-
-        //[Variation(id = 27, Desc = "WriteNode should only write required namespaces", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader27()
-        {
-            string strxml = @"<root xmlns:p1='p1'><p2:child xmlns:p2='p2' /></root>";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            while (xr.Read())
-            {
-                if (xr.LocalName == "child")
-                    break;
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-                xr.Dispose();
-            }
-            Assert.True(CompareReader("<p2:child xmlns:p2='p2' />"));
-        }
-
-        //[Variation(id = 28, Desc = "Reader.WriteNode should only write required namespaces, include xmlns:xml", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader28b()
-        {
-            string strxml = @"<root xmlns:p1='p1'><p2:child xmlns:p2='p2' xmlns:xml='http://www.w3.org/XML/1998/namespace' /></root>";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            while (xr.Read())
-            {
-                if (xr.LocalName == "child")
-                    break;
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-                xr.Dispose();
-            }
-            string exp = (WriterType == WriterType.UnicodeWriter) ? "<p2:child xmlns:p2=\"p2\" />" : "<p2:child xmlns:p2=\"p2\" xmlns:xml='http://www.w3.org/XML/1998/namespace' />";
-            Assert.True(CompareReader(exp));
-        }
-
-        //[Variation(id = 29, Desc = "WriteNode should only write required namespaces, exclude xmlns:xml", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader29()
-        {
-            string strxml = @"<root xmlns:p1='p1' xmlns:xml='http://www.w3.org/XML/1998/namespace'><p2:child xmlns:p2='p2' /></root>";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            while (xr.Read())
-            {
-                if (xr.LocalName == "child")
-                    break;
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-
-                xr.Dispose();
-            }
-            Assert.True(CompareReader("<p2:child xmlns:p2=\"p2\" />"));
-        }
-
-        //[Variation(id = 30, Desc = "WriteNode should only write required namespaces, change default ns at top level", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader30()
-        {
-            string strxml = @"<root xmlns='p1'><child /></root>";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            while (xr.Read())
-            {
-                if (xr.LocalName == "child")
-                    break;
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-                xr.Dispose();
-            }
-            Assert.True(CompareReader("<child xmlns='p1' />"));
-        }
-
-        //[Variation(id = 31, Desc = "WriteNode should only write required namespaces, change default ns at same level", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader31()
-        {
-            string strxml = @"<root xmlns:p1='p1'><child xmlns='p2'/></root>";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            while (xr.Read())
-            {
-                if (xr.LocalName == "child")
-                    break;
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-                xr.Dispose();
-            }
-            Assert.True(CompareReader("<child xmlns='p2' />"));
-        }
-
-        //[Variation(id = 32, Desc = "WriteNode should only write required namespaces, change default ns at both levels", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader32()
-        {
-            string strxml = @"<root xmlns='p1'><child xmlns='p2'/></root>";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            while (xr.Read())
-            {
-                if (xr.LocalName == "child")
-                    break;
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-
-                xr.Dispose();
-            }
-            Assert.True(CompareReader("<child xmlns='p2' />"));
-        }
-
-        //[Variation(id = 33, Desc = "WriteNode should only write required namespaces, change ns uri for same prefix", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader33()
-        {
-            string strxml = @"<p1:root xmlns:p1='p1'><p1:child xmlns:p1='p2'/></p1:root>";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            while (xr.Read())
-            {
-                if (xr.LocalName == "child")
-                    break;
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-                xr.Dispose();
-            }
-            Assert.True(CompareReader("<p1:child xmlns:p1='p2' />"));
-        }
-
-        //[Variation(id = 34, Desc = "WriteNode should only write required namespaces, reuse prefix from top level", Pri = 1)]
-        [Fact]
-        public void writeNode_XmlReader34()
-        {
-            string strxml = @"<root xmlns:p1='p1'><p1:child /></root>";
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            while (xr.Read())
-            {
-                if (xr.LocalName == "child")
-                    break;
-            }
-            using (XmlWriter w = CreateWriter())
-            {
-                w.WriteNode(xr, false);
-                xr.Dispose();
-            }
-            Assert.True(CompareReader("<p1:child xmlns:p1='p1' />"));
-        }
-
-        //[Variation(Desc = "1. XDocument does not format content while Saving", Param = @"<?xml version='1.0'?><?pi?><?pi?>  <shouldbeindented><a>text</a></shouldbeindented><?pi?>")]
-        //[Variation(Desc = "2. XDocument does not format content while Saving", Param = @"<?xml version='1.0'?><?pi?><?pi?>  <shouldbeindented><a>text</a></shouldbeindented><?pi?>")]
-        [Fact]
-        public void writeNode_XmlReader35()
-        {
-            string strxml = (string)CurVariation.Param;
-            CError.WriteLine(strxml);
-            XmlReader xr = CreateReader(new StringReader(strxml));
-            XmlWriterSettings ws = new XmlWriterSettings();
-            ws.ConformanceLevel = (CurVariation.Desc.Contains("1.")) ? ConformanceLevel.Document : ConformanceLevel.Auto;
-            ws.Indent = true;
-            XmlWriter w = CreateWriter(ws);
-            w.WriteNode(xr, false);
-            xr.Dispose();
-            w.Dispose();
-            Assert.True(CompareReader(strxml));
-        }
-
-        //[Variation(Desc = "1.WriteNode with ascii encoding", Param = true)]
-        //[Variation(Desc = "2.WriteNode with ascii encoding", Param = false)]
-        [Fact]
-        public void writeNode_XmlReader36()
-        {
-            string strxml = "<Ro\u00F6t \u00F6=\"\u00F6\" />";
-            string exp = strxml;
-
-            XmlReader xr = CreateReader(new StringReader(strxml));
-
-            XmlWriterSettings ws = new XmlWriterSettings();
-            ws.OmitXmlDeclaration = true;
-            XmlWriter w = CreateWriter(ws);
-            while (!xr.EOF)
-            {
-                w.WriteNode(xr, (bool)CurVariation.Param);
-            }
-            xr.Dispose();
-            w.Dispose();
-            Assert.True((CompareString(exp)));
-        }
-
-        //[Variation(Desc = "WriteNode DTD PUBLIC with identifier", Param = true)]
-        //[Variation(Desc = "WriteNode DTD PUBLIC with identifier", Param = false)]
-        [Fact]
-        public void writeNode_XmlReader37()
-        {
-            string strxml = "<!DOCTYPE root PUBLIC \"\" \"#\"><root/>";
-            string exp = (WriterType == WriterType.UTF8WriterIndent || WriterType == WriterType.UnicodeWriterIndent) ?
-               "<!DOCTYPE root PUBLIC \"\" \"#\"[]>" + nl + "<root />" :
-               "<!DOCTYPE root PUBLIC \"\" \"#\"[]><root />";
-            try
-            {
-                XmlReader xr = CreateReader(new StringReader(strxml));
-                using (XmlWriter w = CreateWriter())
-                {
-                    while (!xr.EOF)
-                    {
-                        w.WriteNode(xr, (bool)CurVariation.Param);
-                    }
-                    xr.Dispose();
-                }
-            }
-            catch (FileNotFoundException e) { CError.WriteLine(e); return; }
-            catch (XmlException e) { CError.WriteLine(e); return; }
-            Assert.True(false);
-        }
-
-        //[Variation(Desc = "WriteNode DTD SYSTEM with identifier", Param = true)]
-        //[Variation(Desc = "WriteNode DTD SYSTEM with identifier", Param = false)]
-        [Fact]
-        public void writeNode_XmlReader38()
-        {
-            string strxml = "<!DOCTYPE root SYSTEM \"#\"><root/>";
-            try
-            {
-                using (XmlReader xr = CreateReader(new StringReader(strxml)))
-                {
-                    using (XmlWriter w = CreateWriter())
-                    {
-                        w.WriteNode(xr, (bool)CurVariation.Param);
-                    }
-                }
-            }
-            catch (XmlException e) { CError.WriteLine(e); return; }
-            Assert.True(false);
-        }
-
-        //[Variation(Desc = "WriteNode DTD SYSTEM with valid surrogate pair", Param = true)]
-        //[Variation(Desc = "WriteNode DTD SYSTEM with valid surrogate pair", Param = false)]
-        [Fact]
-        public void writeNode_XmlReader39()
-        {
-            string strxml = "<!DOCTYPE root SYSTEM \"\uD812\uDD12\"><root/>";
-            string exp = "<!DOCTYPE root SYSTEM \"\uD812\uDD12\"[]><root />";
-            try
-            {
-                using (XmlReader xr = CreateReader(new StringReader(strxml)))
-                {
-                    using (XmlWriter w = CreateWriter())
-                    {
-                        w.WriteNode(xr, (bool)CurVariation.Param);
-                    }
-                }
-            }
-            catch (XmlException e) { CError.WriteLine(e); return; }
-            catch (FileNotFoundException e) { CError.WriteLine(e); return; }
-            if (WriterType == WriterType.CharCheckingWriter)
-            {
-                Assert.True((CompareString(exp)));
-            }
-            Assert.True(false);
-        }
-    }
-
-    //[TestCase(Name = "WriteNode with streaming API ReadValueChunk - COREREADER", Param = "COREREADER")]
-    //[TestCase(Name = "WriteNode with streaming API ReadValueChunk - XMLTEXTREADER", Param = "XMLTEXTREADER")]
-    //[TestCase(Name = "WriteNode with streaming API ReadValueChunk - XMLDOCNODEREADER", Param = "XMLDOCNODEREADER")]
-    //[TestCase(Name = "WriteNode with streaming API ReadValueChunk - XMLDATADOCNODEREADER", Param = "XMLDATADOCNODEREADER")]
-    //[TestCase(Name = "WriteNode with streaming API ReadValueChunk - XPATHDOCNAVIGATORREADER", Param = "XPATHDOCNAVIGATORREADER")]
-    //[TestCase(Name = "WriteNode with streaming API ReadValueChunk - XMLDOCNAVIGATORREADER", Param = "XMLDOCNAVIGATORREADER")]
-    //[TestCase(Name = "WriteNode with streaming API ReadValueChunk - XMLDATADOCNAVIGATORREADER", Param = "XMLDATADOCNAVIGATORREADER")]
-    public partial class TCWriteNode_With_ReadValueChunk : ReaderParamTestCase
-    {
-        /* Buffer size is 1024 chars in XmlWriter */
-        private XmlReader CreateReader(int size)
-        {
-            StringBuilder sb = new StringBuilder();
-            sb.Append("<root>");
-            for (int i = 0; i < size; i++)
-            {
-                sb.Append("A");
-            }
-            sb.Append("</root>");
-
-            StringReader sr = new StringReader(sb.ToString());
-            XmlReader r = base.CreateReader(sr);
-            return r;
-        }
-
-        //[Variation(id = 1, Desc = "Input XML in utf8 encoding, text node has 1K-1 chars", Pri = 0, Param = "1023")]
-        //[Variation(id = 2, Desc = "Input XML in utf8 encoding, text node has 1K chars", Pri = 0, Param = "1024")]
-        //[Variation(id = 3, Desc = "Input XML in utf8 encoding, text node has 1K+1 chars", Pri = 0, Param = "1025")]
-        //[Variation(id = 4, Desc = "Input XML in utf8 encoding, text node has 2K chars", Pri = 0, Param = "2048")]
-        //[Variation(id = 5, Desc = "Input XML in utf8 encoding, text node has 4K chars", Pri = 0, Param = "4096")]
-        [Fact]
-        public void writeNode_1()
-        {
-            int size = Int32.Parse(CurVariation.Param.ToString());
-            using (XmlReader r = this.CreateReader(size))
-            {
-                using (XmlWriter w = CreateWriter())
-                {
-                    while (r.Read())
-                    {
-                        w.WriteNode(r, false);
-                    }
-                }
-            }
-
-            switch (size)
-            {
-                case 1023:
-                    Assert.True(CompareBaseline("textnode_1K-1_utf8.xml"));
-                    break;
-                case 1024:
-                    Assert.True(CompareBaseline("textnode_1K_utf8.xml"));
-                    break;
-                case 1025:
-                    Assert.True(CompareBaseline("textnode_1K+1_utf8.xml"));
-                    break;
-                case 2048:
-                    Assert.True(CompareBaseline("textnode_2K_utf8.xml"));
-                    break;
-                case 4096:
-                    Assert.True(CompareBaseline("textnode_4K_utf8.xml"));
-                    break;
-            }
-            CError.WriteLine("Error");
-            Assert.True(false);
-        }
-
-        //[Variation(id = 6, Desc = "Input XML in unicode encoding, text node has 1K-1 chars", Pri = 0, Param = "1023")]
-        //[Variation(id = 7, Desc = "Input XML in unicode encoding, text node has 1K chars", Pri = 0, Param = "1024")]
-        //[Variation(id = 8, Desc = "Input XML in unicode encoding, text node has 1K+1 chars", Pri = 0, Param = "1025")]
-        //[Variation(id = 9, Desc = "Input XML in unicode encoding, text node has 2K chars", Pri = 0, Param = "2048")]
-        //[Variation(id = 10, Desc = "Input XML in unicode encoding, text node has 4K chars", Pri = 0, Param = "4096")]
-        [Fact]
-        public void writeNode_2()
-        {
-            int size = Int32.Parse(CurVariation.Param.ToString());
-            using (XmlReader r = this.CreateReader(size))
-            {
-                using (XmlWriter w = CreateWriter())
-                {
-                    while (r.Read())
-                    {
-                        w.WriteNode(r, false);
-                    }
-                }
-            }
-
-            switch (size)
-            {
-                case 1023:
-                    Assert.True(CompareBaseline("textnode_1K-1_unicode.xml"));
-                    break;
-                case 1024:
-                    Assert.True(CompareBaseline("textnode_1K_unicode.xml"));
-                    break;
-                case 1025:
-                    Assert.True(CompareBaseline("textnode_1K+1_unicode.xml"));
-                    break;
-                case 2048:
-                    Assert.True(CompareBaseline("textnode_2K_unicode.xml"));
-                    break;
-                case 4096:
-                    Assert.True(CompareBaseline("textnode_4K_unicode.xml"));
-                    break;
-            }
-            CError.WriteLine("Error");
-            Assert.True(false);
-        }
-
-
-        //[Variation(id = 11, Desc = "Trailing surrogate pair", Pri = 1)]
-        [Fact]
-        public void writeNode_3()
-        {
-            StringBuilder sb = new StringBuilder();
-            sb.Append("<root>");
-            for (int i = 0; i < 1022; i++)
-            {
-                sb.Append("A");
-            }
-            sb.Append("&#x10000;");
-            sb.Append("</root>");
-
-            StringReader sr = new StringReader(sb.ToString());
-            using (XmlReader r = ReaderHelper.Create(sr))
-            {
-                using (XmlWriter w = CreateWriter())
-                {
-                    while (r.Read())
-                    {
-                        w.WriteNode(r, false);
-                    }
-                }
-            }
-            Assert.True(CompareBaseline("trailing_surrogate_1K.xml"));
-        }
-
-        //[Variation(id = 12, Desc = "Leading surrogate pair", Pri = 1)]
-        [Fact]
-        public void writeNode_4()
-        {
-            StringBuilder sb = new StringBuilder();
-            sb.Append("<root>");
-            sb.Append("&#x10000;");
-            for (int i = 0; i < 1022; i++)
-            {
-                sb.Append("A");
-            }
-            sb.Append("</root>");
-
-            StringReader sr = new StringReader(sb.ToString());
-            XmlReader r = ReaderHelper.Create(sr);
-            using (XmlWriter w = CreateWriter())
-            {
-                while (r.Read())
-                {
-                    w.WriteNode(r, false);
-                }
-            }
-            r.Dispose();
-
-            Assert.True(CompareBaseline("leading_surrogate_1K.xml"));
-        }
-
-        //[Variation(id = 13, Desc = "Split surrogate pair across 1K buffer boundary", Pri = 1)]
-        [Fact]
-        public void writeNode_5()
-        {
-            StringBuilder sb = new StringBuilder();
-            sb.Append("<root>");
-
-            for (int i = 0; i < 1023; i++)
-            {
-                sb.Append("A");
-            }
-            sb.Append("&#x10000;");
-            sb.Append("</root>");
-
-            StringReader sr = new StringReader(sb.ToString());
-            XmlReader r = ReaderHelper.Create(sr);
-            using (XmlWriter w = CreateWriter())
-            {
-                while (r.Read())
-                {
-                    w.WriteNode(r, false);
-                }
-            }
-            r.Dispose();
-
-            Assert.True(CompareBaseline("split_surrogate_1K.xml"));
-        }
-    }
-
     //[TestCase(Name = "WriteFullEndElement")]
-    public partial class TCFullEndElement : XmlWriterTestCaseBase
+    public class TCFullEndElement : XmlWriterTestCaseBase
     {
         //[Variation(id = 1, Desc = "Sanity test for WriteFullEndElement()", Pri = 0)]
         [Fact]
